requirements.txt
openai-agents>=0.6.0
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
